+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
 ---
-### Version [0.0.0]:
+### Version [0.1.0] --- June 7th, 2025:
+- Full datasets can now be loaded via icegraph.dataset.Data.from_config().
+- Added configuration handling via icegraph.config.Config.
+- Significantly improved internal documentation.
+- Condensed user configs to one file for usability, internal configs are separate.
+- Slightly optimized caching for faster repeated conversions via icegraph.cache.I3ConversionCache.
+- Other minor changes.
+
+### Version [0.0.0] --- June 6th, 2025:
 - Implementation of semantic versioning. See https://semver.org/.